+++ item_book.rb
@@ -1,0 +1,6 @@
+class Book < Item
+  def initialize(args)
+    super(args)
+    @exempt = true
+  end
+end
