+++ item_cosmetic.rb
@@ -1,0 +1,5 @@
+class Cosmetic < Item
+  def initialize(args)
+    super(args)
+  end
+end
