+++ item_food.rb
@@ -1,0 +1,9 @@
+class Food < Item
+  def initialize(args)
+    super(args)
+    @exempt = true
+  end
+
+  
+
+end
