+++ item_media.rb
@@ -1,0 +1,5 @@
+class Media < Item
+  def initialize(args)
+    super(args)
+  end
+end
