+++ item_medical.rb
@@ -1,0 +1,6 @@
+class Medical < Item
+  def initialize(args)
+    super(args)
+    @exempt = true
+  end
+end
